Docker update for rocBLAS, HCC, and HIP to 2.1 (#1606)
* Updated the requirements with newer HIP, HCC, rocBLAS, and clang-ocl

* Fixed boost issue.

* Fixed tidy issues across the lib.

* Formatting.

* Update hip and rocblas version.

* Fixed remaining tidy issues from formatting.

* Formatting.

* More tidy issues.

* Float conversion.

* NOLINT on stubborn coversion.

* ktime to float

* Remove unneeded NOLINT

* Removed unneeded semi-colon

* Drive updates.

* Switched to rocBLAS 2.3 RC commit hash.

* Use size_t for fusion op_map.

+++ src/ocl/lrn_ocl.cpp
@@ -75,7 +75,7 @@ miopenStatus_t LRNDescriptor::Forward(Handle& handle,
     construct_params.setBotDescFromMLDesc(xDesc);
 
     int norm_reg     = GetMode();
-    int local_area   = GetN();
+    int local_area   = static_cast<int>(GetN());
     double lrn_alpha = GetAlpha();
     double lrn_beta  = GetBeta();
     double lrn_K     = GetK();
@@ -250,7 +250,7 @@ miopenStatus_t LRNDescriptor::Backward(Handle& handle,
 
     int norm_reg = GetMode();
 
-    int local_area = GetN();
+    int local_area = static_cast<int>(GetN());
 
     double lrn_alpha = GetAlpha();
     double lrn_beta  = GetBeta();
